created abstract url builder and updated the hidden inputs to reflect api url keys
